Run raw-block Rust unit tests in CI
