+++ rust/raw_block/src/lib.rs
@@ -546,34 +546,6 @@ fn placement_id_to_u16(pid: i32) -> PyResult<u16> {
     u16::try_from(pid).map_err(|_| PyValueError::new_err("placement_id must be in range 1..=65535"))
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn check_nvme_ioctl_result_accepts_success() {
-        assert!(check_nvme_ioctl_result(0, "NVMe ioctl failed").is_ok());
-    }
-
-    #[test]
-    fn check_nvme_ioctl_result_rejects_nvme_status() {
-        assert!(check_nvme_ioctl_result(1, "NVMe ioctl failed").is_err());
-    }
-
-    #[test]
-    fn placement_id_to_u16_accepts_valid_bounds() {
-        assert_eq!(placement_id_to_u16(1).unwrap(), 1);
-        assert_eq!(placement_id_to_u16(65535).unwrap(), 65535);
-    }
-
-    #[test]
-    fn placement_id_to_u16_rejects_reserved_and_out_of_range_values() {
-        assert!(placement_id_to_u16(0).is_err());
-        assert!(placement_id_to_u16(-1).is_err());
-        assert!(placement_id_to_u16(65536).is_err());
-    }
-}
-
 /// Prepare NVMe uring command for read/write operations
 #[allow(clippy::too_many_arguments)]
 fn nvme_uring_cmd_prep(
@@ -3352,3 +3324,6 @@ fn lmcache_rust_raw_block_io(_py: Python, m: &Bound<'_, PyModule>) -> PyResult<(
     m.add_class::<RawBlockDevice>()?;
     Ok(())
 }
+
+#[cfg(test)]
+mod tests;
